Metadata now recognizes the block ranges of terse functions (#543)
* Metadata now recognizes the block ranges of terse functions

While working on the structure PR, I noticed that the indexer was
returning both definitions and references for functions defined
tersely, that is without a do/end. For example:

```elixir
def add_one(arg), do: arg + 1
```

Somewhat surprisingly, this is because the metadata location function
was not correctly recognizing this as a block, and was returning a
single position. The block code thought this was just an expression
and they weren't properly registered as blocks, but the indexer
correctly recognized them as definitions.

This PR improves detection of blocks by the metadata to recognize
multiple types of blocks, whether or not they're terse.

Fixes #540

* Updated sourceror, removed workaround for its bug