Struct discovery now uses the index (#582)
* Struct discovery now uses the index

Thehe last thing that used the forced build was struct discovery. Now
that we index struct definitions, we can instead leverage the already
built index, and not have to force compile projects when they start.

This has a tremendous impact on startup time. Lexical would take 11
seconds to compile on my system, and now it compiles in 1.9
seconds. The index is ready nearly instantly after that.

This does mean that this is a bit of a rubicon commit, after this, we
rely on the index for core functionality, and we can't disable it. As
such, I've removed the environment variable, and the feature flag.