Removed select_all() from backend in favor of a reduce function
There are several places where we're pulling all the entries from the
backend, processing them and throwing them out. This creates memory
pressure on every process involved in handing the entries.

Instead of that, it makes more sense to expose a reduce function
inside the backend that allows you to process the entries however you
want, and only the data you care about is pulled through the
processes.

This saves around 250mb running the lexical project on my laptop.