Stopped indexing references in deps (#591)
We were indexing references for modules and functions in
dependencies. This made the index much larger than it needed to be,
and would cloud "find references" with references that your project
didn't make. I could see it being useful in niche situations, but the
default should be to not index them.

On lexical, this halves memory usage by the VM

In the future, this could be a per-project setting

Fixes #590