Added variable extractor and ability to customize which extractors run
Added an extractor that pulls out variables and references from
AST. It's likely incomplete, but we can add missed cases as we go.

I also added the ability to customize which extractors run during an
index. The idea here is to have a separate indexing run after a file
is compiled, so we can pull out variables from it and power go to
definition for variables without polluting the index with things that
we won't use or need.