Fixed bug where blocks weren't popped correctly
While traversing the AST, it's pretty common to not get the position
updated consistently. Because of this, it's possible that there might
be more than one block that needs to be popped at a time.

We were also not attempting to pop blocks when an expression occurred,
which would cause block mismatches in an expression that occurred
after a block ended.