chore: update FOD hash
It seems that there is still issue with autoupdater. It seems that
`main` is completely locked for direct commits, which mean that GitHub
Bot cannot update the hash automatically. I need to review the
possibility to automatically create PR from the GHA to handle that. But
for now we need to update hash manually.