Added ranges to aliases
When doing refactor operations like organize imports, it would be
helpful if we could get the entire range for an alias, because if we do,
replacing the alias is a trivial matter of replacing the range with an
empty string. Previously, aliases just had a line number, but this won't
work for aliases with blocks.