Fix conversion of UTF-8 positions to UTF-16 code units (#719)
Internally, Lexical treats lines and columns the same way that Elixir does, which are 1-based and, in the case of columns/characters, based on UTF-8 graphemes. Before talking to an LS client, these are converted into positions based on UTF-16 code units, which is what LSP speaks.

There were two bugs in this conversion that are fixed here:

- The conversion assumed that Lexical's character positions were UTF-8 code units (bytes), instead of graphemes, so the wrong conversion was being used (UTF-8 code unit to UTF-16 code unit).
- The correct conversion (UTF-8 character position to UTF-16 code unit) was not properly handling certain multi-code-point graphemes.

While the corrected code is noticeably slower in benchmarks, the impact should be minimal because a simplified conversion can be used for lines that are pure ASCII, which is a large majority of source files.