Proxy for expensive API requests
When doing complicated refactors, the language server will often spam
remote_control with a ton of operations when the refactor is going on.
These operations will slow down both the editor and remote_control so
that it's not particularly usable. In order to mitigate this, certain
operations, like project compilation, can be buffered and completed when
the refactor has completed.

This PR implements a proxy module that buffers these complicated
operations, and monitors the calling process. When that process exits,
the buffered operations are emitted in remote_control, allowing them to
complete without disrupting the refactor.