Fixed deadlocks
In testing, I found two deadlocks. The first was that the build server
was using RemoteControl.call to invoke itself rather than
`GenServer.call`. This created an additional lock, and was
unnecessary, because it's running locally.

The second was more insidious. The proxy module made a bunch of
separate calls to other modules, which in turn woud call
`Build.with_lock`, which would then be run inside the proxy's
process. This would cause fairly frequent locks in the proxy process.
Moving these calls into tasks, fixes things, but creates the need for
an additional `draining` state of the proxy, which it enteres if there
are in-flight requests when `start_buffering` is called.