Sorted bang functions after non-bang variants
Bang functions were being sorted after non-bang functions because ! is
a very low character (the only other printable character smaller is a
space). This means that "003_my_function!:002" would be sorted before
"003_my_function:002", which isn't really what we want. Changing the
separator between the name and arity to space fixes the problem.