[feat] Loosen find_references
Prior, find references would take the arity into account. This was
because in elixir, a functin is defined by its name and its
arity. However in practice, you more than not want to see a all the
calls of a function with a given name rather than a specific
head. This is especially true for functions with default arguments.

As a result, this change looks for functions with a given name and any arity.