Performance oriented fixes. (#841)
* Perf: Precalculate contexts

We were calculating the context of the cursor once per suggestion,
this would take a couple dozen milliseconds, but for a module that had
a lot of functions, this would add up. Precalculating them saved a lot
of time.

* Use new Mix.Tasks.Format

Jose added a function that allows us to get a formatter without having
to lock on the build. This reduces the formatting time significantly,
and we no longer have that frustrating freeze on save.

---------

Co-authored-by: scohen <scohen@users.noreply.github.com>